Implemented basic thread per RecordAndSubmitTask
